Add files via upload
This script uses a Random Forest Regressor to predict salaries based on years of experience. It trains the model on the given data and visualizes the predictions against the original data points. Additionally, it provides predictions for given years of experience. The script calculates the R2 score to evaluate the Random Forest model's performance in predicting salaries accurately.