Add files via upload
This Python script is designed to assess the performance of logistic regression models on the Wine dataset, employing two distinct dimensionality reduction techniques: Principal Component Analysis (PCA) and Linear Discriminant Analysis (LDA). Initially, the dataset is imported and split into features (X) and labels (y). Following this, the data is partitioned into training and testing sets, and standard scaling is applied to ensure uniformity across features.

Subsequently, PCA is implemented to reduce the dataset's dimensionality to two principal components. The PCA model is fitted to the training data, and both the training and testing sets are transformed accordingly. Logistic regression models are then trained using the original data and the PCA-transformed data.

To evaluate model performance, predictions are made using the trained logistic regression models on the testing data. Additionally, confusion matrices are computed to provide insights into the models' classification accuracy. These matrices are generated for the logistic regression model without PCA, with PCA, and using LDA on the original data. By comparing these results, the effectiveness of dimensionality reduction techniques in enhancing classification performance can be assessed.